test(datastore): move test utils to directory
Remove test utils from coverage path
