+++ packages/datastore/__tests__/__utils__/commonAdapterTests.ts
@@ -2,7 +2,7 @@ import {
 	DataStore as DataStoreType,
 	PersistentModelConstructor,
 	initSchema as initSchemaType,
-} from '../src/';
+} from '../../src/';
 
 import {
 	pause,
